reusable has-statement filter
