+++ c7n/resources/s3.py
@@ -54,6 +54,7 @@ from c7n.exceptions import PolicyValidationError, PolicyExecutionError
 from c7n.filters import (
     FilterRegistry, Filter, CrossAccountAccessFilter, MetricsFilter,
     ValueFilter)
+import c7n.filters.policystatement as polstmt_filter
 from c7n.manager import resources
 from c7n.output import NullBlobOutput
 from c7n import query
@@ -811,93 +812,15 @@ class BucketFinding(PostFinding):
         return filter_empty(resource)
 
 
-@filters.register('has-statement')
-class HasStatementFilter(BucketFilterBase):
-    """Find buckets with set of policy statements.
-
-    :example:
-
-    .. code-block:: yaml
-
-            policies:
-              - name: s3-bucket-has-statement
-                resource: s3
-                filters:
-                  - type: has-statement
-                    statement_ids:
-                      - RequiredEncryptedPutObject
-
-
-            policies:
-              - name: s3-public-policy
-                resource: s3
-                filters:
-                  - type: has-statement
-                    statements:
-                      - Effect: Allow
-                        Action: 's3:*'
-                        Principal: '*'
-    """
-    schema = type_schema(
-        'has-statement',
-        statement_ids={'type': 'array', 'items': {'type': 'string'}},
-        statements={
-            'type': 'array',
-            'items': {
-                'type': 'object',
-                'properties': {
-                    'Sid': {'type': 'string'},
-                    'Effect': {'type': 'string', 'enum': ['Allow', 'Deny']},
-                    'Principal': {'anyOf': [
-                        {'type': 'string'},
-                        {'type': 'object'}, {'type': 'array'}]},
-                    'NotPrincipal': {
-                        'anyOf': [{'type': 'object'}, {'type': 'array'}]},
-                    'Action': {
-                        'anyOf': [{'type': 'string'}, {'type': 'array'}]},
-                    'NotAction': {
-                        'anyOf': [{'type': 'string'}, {'type': 'array'}]},
-                    'Resource': {
-                        'anyOf': [{'type': 'string'}, {'type': 'array'}]},
-                    'NotResource': {
-                        'anyOf': [{'type': 'string'}, {'type': 'array'}]},
-                    'Condition': {'type': 'object'}
-                },
-                'required': ['Effect']
-            }
-        })
-
-    def process(self, buckets, event=None):
-        return list(filter(None, map(self.process_bucket, buckets)))
-
-    def process_bucket(self, b):
-        p = b.get('Policy')
-        if p is None:
-            return None
-        p = json.loads(p)
-
-        required = list(self.data.get('statement_ids', []))
-        statements = p.get('Statement', [])
-        for s in list(statements):
-            if s.get('Sid') in required:
-                required.remove(s['Sid'])
-
-        required_statements = format_string_values(list(self.data.get('statements', [])),
-                                                   **self.get_std_format_args(b))
-        for required_statement in required_statements:
-            for statement in statements:
-                found = 0
-                for key, value in required_statement.items():
-                    if key in statement and value == statement[key]:
-                        found += 1
-                if found and found == len(required_statement):
-                    required_statements.remove(required_statement)
-                    break
-
-        if (self.data.get('statement_ids', []) and not required) or \
-           (self.data.get('statements', []) and not required_statements):
-            return b
-        return None
+@S3.filter_registry.register('has-statement')
+class HasStatementFilter(polstmt_filter.HasStatementFilter, BucketFilterBase):
+    def get_std_format_args(self, bucket):
+        return {
+            'account_id': self.manager.config.account_id,
+            'region': self.manager.config.region,
+            'bucket_name': bucket['Name'],
+            'bucket_region': get_region(bucket)
+        }
 
 
 ENCRYPTION_STATEMENT_GLOB = {
