+++ c7n/resources/sns.py
@@ -3,8 +3,9 @@
 import json
 
 from c7n.actions import RemovePolicyBase, ModifyPolicyBase, BaseAction
-from c7n.filters import CrossAccountAccessFilter, PolicyChecker, Filter
+from c7n.filters import CrossAccountAccessFilter, PolicyChecker
 from c7n.filters.kms import KmsRelatedFilter
+import c7n.filters.policystatement as polstmt_filter
 from c7n.manager import resources
 from c7n.query import ConfigSource, DescribeSource, QueryResourceManager, TypeInfo
 from c7n.resolver import ValuesFrom
@@ -78,93 +79,11 @@ class SNSPostFinding(PostFinding):
 
 
 @SNS.filter_registry.register('has-statement')
-class HasStatementFilter(Filter):
-    """Find SNS topics with set of matching access policy statements.
-
-    :example:
-
-    .. code-block:: yaml
-
-            policies:
-              - name: sns-check-statement-id
-                resource: sns
-                filters:
-                  - type: has-statement
-                    statement_ids:
-                      - BlockNonSSL
-
-            policies:
-              - name: sns-check-block-non-ssl
-                resource: sns
-                filters:
-                  - type: has-statement
-                    statements:
-                      - Effect: Deny
-                        Action: 'SNS:Publish'
-                        Principal: '*'
-                        Condition:
-                            Bool:
-                                "aws:SecureTransport": "false"
-    """
-    schema = type_schema(
-        'has-statement',
-        statement_ids={'type': 'array', 'items': {'type': 'string'}},
-        statements={
-            'type': 'array',
-            'items': {
-                'type': 'object',
-                'properties': {
-                    'Sid': {'type': 'string'},
-                    'Effect': {'type': 'string', 'enum': ['Allow', 'Deny']},
-                    'Principal': {'anyOf': [
-                        {'type': 'string'},
-                        {'type': 'object'}, {'type': 'array'}]},
-                    'NotPrincipal': {
-                        'anyOf': [{'type': 'object'}, {'type': 'array'}]},
-                    'Action': {
-                        'anyOf': [{'type': 'string'}, {'type': 'array'}]},
-                    'NotAction': {
-                        'anyOf': [{'type': 'string'}, {'type': 'array'}]},
-                    'Resource': {
-                        'anyOf': [{'type': 'string'}, {'type': 'array'}]},
-                    'NotResource': {
-                        'anyOf': [{'type': 'string'}, {'type': 'array'}]},
-                    'Condition': {'type': 'object'}
-                },
-                'required': ['Effect']
-            }
-        })
-
-    def process(self, topics, event=None):
-        return list(filter(None, map(self.process_topic, topics)))
-
-    def process_topic(self, t):
-        p = t.get('Policy')
-        if p is None:
-            return None
-        p = json.loads(p)
-
-        required = list(self.data.get('statement_ids', []))
-        statements = p.get('Statement', [])
-        for s in list(statements):
-            if s.get('Sid') in required:
-                required.remove(s['Sid'])
-
-        required_statements = list(self.data.get('statements', []))
-        for required_statement in required_statements:
-            for statement in statements:
-                found = 0
-                for key, value in required_statement.items():
-                    if key in statement and value == statement[key]:
-                        found += 1
-                if found and found == len(required_statement):
-                    required_statements.remove(required_statement)
-                    break
-
-        if (self.data.get('statement_ids', []) and not required) or \
-           (self.data.get('statements', []) and not required_statements):
-            return t
-        return None
+class HasStatementFilter(polstmt_filter.HasStatementFilter):
+    def get_std_format_args(self, topic):
+        return {
+            'topic_arn': topic['TopicArn']
+        }
 
 
 @SNS.action_registry.register('tag')
